add source and move Manga4Life to tempalte parser and fix url Duniakomik
