+++ src/main/kotlin/org/koitharu/kotatsu/parsers/site/mangareader/id/Duniakomik.kt
@@ -8,7 +8,7 @@ import org.koitharu.kotatsu.parsers.site.mangareader.MangaReaderParser
 
 @MangaSourceParser("DUNIAKOMIK", "Duniakomik", "id", ContentType.HENTAI)
 internal class Duniakomik(context: MangaLoaderContext) :
-	MangaReaderParser(context, MangaSource.DUNIAKOMIK, "duniakomik.id", pageSize = 12, searchPageSize = 12) {
+	MangaReaderParser(context, MangaSource.DUNIAKOMIK, "duniakomik.org", pageSize = 12, searchPageSize = 12) {
 
 	override val datePattern = "MMM d, yyyy"
 }
