+++ src/main/kotlin/org/koitharu/kotatsu/parsers/site/nepnep/en/Manga4Life.kt
@@ -1,0 +1,10 @@
+package org.koitharu.kotatsu.parsers.site.nepnep.en
+
+import org.koitharu.kotatsu.parsers.MangaLoaderContext
+import org.koitharu.kotatsu.parsers.MangaSourceParser
+import org.koitharu.kotatsu.parsers.model.MangaSource
+import org.koitharu.kotatsu.parsers.site.nepnep.NepnepParser
+
+@MangaSourceParser("MANGA4LIFE", "Manga4Life", "en")
+internal class Manga4Life(context: MangaLoaderContext) :
+	NepnepParser(context, MangaSource.MANGA4LIFE, "manga4life.com")
